Phase-II/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
python-dotenv>=1.0.0
